EditorModel: add getUpdatedData
Remove getUpdatedDataFromGrid and getUpdatedDataFromEditableGrid
